Rename initoverlayfs-setup to initoverlayfs
This binary is going to have two modes:

initoverlayfs --systemd

which is the recommended one. And:

initoverlayfs

where we use initoverlayfs's built-in init system, this is for use-cases
where we require high levels of optimization, such as in the Automotive
use case, this requires storage drivers to be built directly into the
kernel so the kernel can do storage initialization, this built-in
initoverlayfs init system will not do udev type things, at least
initially.
